backend+webview/requirements.txt
pymongo==2.7.1
parse==1.6.6
google-api-python-client==1.4.0
Flask==0.10.1
Pillow==2.8.1
requests==1.2.3
